+++ app/src/main/java/com/leon/estimate/Tables/Zarib.java
@@ -1,18 +1,18 @@
 package com.leon.estimate.Tables;
 
 public class Zarib {
-    public int id;
-    public int zoneId;
-    public int maskkoniZ1;
-    public int tejariZ1;
-    public int edariDolatiZ1;
-    public int khadamatiZ1;
-    public int sanatiZ1;
-    public int sayerZ1;
-    public int maskooniZ2;
-    public int tejariZ2;
-    public int edariDolatiZ2;
-    public int khadamatiZ2;
-    public int sanatiZ2;
-    public int sayerZ2;
+    public double id;
+    public double zoneId;
+    public double maskkoniZ1;
+    public double tejariZ1;
+    public double edariDolatiZ1;
+    public double khadamatiZ1;
+    public double sanatiZ1;
+    public double sayerZ1;
+    public double maskooniZ2;
+    public double tejariZ2;
+    public double edariDolatiZ2;
+    public double khadamatiZ2;
+    public double sanatiZ2;
+    public double sayerZ2;
 }
